Padrão Abstract Factory. Refatoração das classes de escudo. Diferentes ataques e poderes agora atuam como modificadores de ataque.

+++ Strategy/Ataque.java
@@ -2,7 +2,7 @@ package Strategy;
 
 public abstract class Ataque
 {
-    public abstract int ataque();
+    public abstract double ataque(); // retorna o modificar de ataque (x0.5 x1.0 ou x1.5)
     
     // métodos usados pela máquina de estados de modo a evitar a perda de decoradores
     public Ataque getProx()
